feat: 게시판 CRUD User 검증 단위 테스트 및 로직 구현

+++ src/main/java/mi/board/domain/board/service/BoardService.java
@@ -3,12 +3,16 @@ package mi.board.domain.board.service;
 import lombok.RequiredArgsConstructor;
 import mi.board.domain.board.dto.BoardRequestDto;
 import mi.board.domain.board.entity.Board;
+import mi.board.domain.board.exception.BoardNotFoundException;
+import mi.board.domain.board.exception.UnauthorizedAccessException;
 import mi.board.domain.board.repository.BoardRepository;
+import mi.board.domain.user.entity.User;
+import mi.board.global.security.userdetails.UserDetailsImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.NoSuchElementException;
+
 
 @Service
 @RequiredArgsConstructor
@@ -16,8 +20,10 @@ public class BoardService {
     private final BoardRepository boardRepository;
 
     @Transactional
-    public Board createBoard(BoardRequestDto boardRequestDto) {
-        Board board = boardRequestDto.toEntity();
+    public Board createBoard(BoardRequestDto boardRequestDto, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
+
+        Board board = boardRequestDto.toEntity(user);
         return boardRepository.save(board);
     }
 
@@ -29,23 +35,36 @@ public class BoardService {
     @Transactional(readOnly = true)
     public Board getBoardById(Long boardId) {
         return boardRepository.findById(boardId).orElseThrow(
-                () -> new NoSuchElementException("찾을 수 없는 게시물입니다.")
-        );
+                BoardNotFoundException::new);
+
     }
 
     @Transactional
-    public Board updateBoard(Long boardId, BoardRequestDto newBoardRequestDto) {
+    public Board updateBoard(Long boardId, BoardRequestDto newBoardRequestDto, UserDetailsImpl userDetails) {
         Board board = boardRepository.findById(boardId).orElseThrow(
-                () -> new NoSuchElementException("찾을 수 없는 게시물입니다.")
-        );
+                BoardNotFoundException::new);
+
+        validateUserAuthorization(userDetails.getUser(), board);
+
         board.update(newBoardRequestDto);
-        return boardRepository.save(board);
+        return board;
     }
 
-    public void deleteBoardById(Long boardId) {
+    @Transactional
+    public void deleteBoardById(Long boardId, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
+
         Board board = boardRepository.findById(boardId).orElseThrow(
-                () -> new NoSuchElementException("찾을 수 없는 게시물입니다.")
-        );
+                BoardNotFoundException::new);
+
+        validateUserAuthorization(userDetails.getUser(), board);
+
         boardRepository.deleteById(boardId);
+    }
+
+    private void validateUserAuthorization(User user, Board board) {
+        if (!user.getUsername().equals(board.getUser().getUsername())) {
+            throw new UnauthorizedAccessException();
+        }
     }
 }
